+++ src/app/zerodha/_redux/sagas/Bnance/socket.ts
@@ -1,9 +1,11 @@
+"use client";
+import { Middleware } from "redux";
 import {
   updateBinanceWSStats,
   updateBinanceWSSubsriptions,
 } from "../../Slices/BinanceWSStats";
 import { updateLivestream } from "../../Slices/Livestream";
-import { store } from "../../store";
+import { RootState, store } from "../../store";
 
 export type TsymbolTrade = {
   e: string;
@@ -24,14 +26,14 @@ export type Twsbinance = {
   params: string[];
   id: number;
 };
-const setupSocket = (dispatch: typeof store.dispatch, url: string) => {
+const setupSocket = (url: string) => {
   const socket = new WebSocket(url);
   socket.onclose = () => {
-    dispatch(updateBinanceWSStats(false));
+    store.dispatch(updateBinanceWSStats(false));
   };
   socket.onopen = () => {
     console.log("redux-saga connected websocket to binance");
-    dispatch(updateBinanceWSStats(true));
+    store.dispatch(updateBinanceWSStats(true));
   };
   socket.onmessage = (event) => {
     const data = JSON.parse(event.data as string) as
@@ -41,13 +43,72 @@ const setupSocket = (dispatch: typeof store.dispatch, url: string) => {
       console.log("redux-saga recieved message", data);
       if (data.result) {
         console.log(typeof data.result, data.result);
-        dispatch(updateBinanceWSSubsriptions(data.result));
+        store.dispatch(updateBinanceWSSubsriptions(data.result));
       }
     } else {
-      dispatch(updateLivestream({ curPrice: data.p, symbol: data.s }));
+      store.dispatch(updateLivestream({ curPrice: data.p, symbol: data.s }));
     }
   };
-  return socket;
+  function sendWSBinanceMessage(message: Twsbinance) {
+    while (socket.readyState !== 1) {}
+    console.log("socket sendingmessage", message);
+    socket.send(JSON.stringify(message));
+  }
+
+  function socketSend(payload: Twsbinance) {
+    payload.params = payload.params
+      .map((item) => {
+        const [first, second] = item.split("@");
+        return first?.toLowerCase() + "@" + second;
+      })
+      .filter((i) => i !== "@trade");
+
+    if (payload.params.length) sendWSBinanceMessage(payload);
+  }
+  const subUnsubMddleware: Middleware = (store) => (next) => (action) => {
+    const { type } = JSON.parse(JSON.stringify(action)) as { type: string };
+    let unsubparams: string[] = [];
+    if (
+      type === "watchListType/updateWatchListNo" ||
+      type === "watchListType/updateWatchList" ||
+      type === "headerPinType/updateHeaderPin"
+    ) {
+      const oldState = store.getState() as RootState;
+      unsubparams = [
+        ...(oldState.watchList.List[oldState.watchList.ListNo] ?? []),
+        oldState.headerPin.Pin0,
+        oldState.headerPin.Pin1,
+      ].map((i) => i + "@trade");
+    }
+    next(action);
+
+    if (
+      type === "watchListType/updateWatchListNo" ||
+      type === "watchListType/updateWatchList" ||
+      type === "headerPinType/updateHeaderPin"
+    ) {
+      const newState = store.getState() as RootState;
+      const subparams = [
+        ...(newState.watchList.List[newState.watchList.ListNo] ?? []),
+        newState.headerPin.Pin0,
+        newState.headerPin.Pin1,
+      ].map((i) => i + "@trade");
+
+      const common = subparams.filter((x) => unsubparams.includes(x));
+      socketSend({
+        method: "SUBSCRIBE",
+        params: subparams.filter((i) => !common.includes(i)),
+        id: 1,
+      });
+      socketSend({
+        method: "UNSUBSCRIBE",
+        params: unsubparams.filter((i) => !common.includes(i)),
+        id: 1,
+      });
+      socket.send(JSON.stringify({ method: "LIST_SUBSCRIPTIONS", id: 3 }));
+    }
+  };
+  return subUnsubMddleware;
 };
 
 export default setupSocket;
